pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi-template"
version = "1.0.0"
description = "A template for FastAPI"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "sqlalchemy",
    "alembic",
    "asyncpg",
    "psycopg2-binary",
    "redis",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "python-dotenv",
    "pydantic",
    "pydantic-settings",
    "email-validator",
    "loguru",
    "hatchling",
    "mypy",
    "ruff",
]



[project.scripts]
dev = "src.settings.run:dev_command"
prod = "src.settings.run:prod_command"
format = "src.settings.run:ruff_fix"
lint = "src.settings.run:ruff_lint"
makemigrations = "src.settings.run:makemigrations"
migrate = "src.settings.run:migrate"
createsuperuser = "scripts.create_superuser:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
package = true

[tool.ruff]
target-version = "py311"
line-length = 120
fix = true
lint.select = [
    "F",
    "E",
    "W",
    "C",
    "UP",
    "I",
]
lint.ignore = [
    "E402",
    "UP006",
    "UP007",
    "E741",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401",
    "F403",
]

[tool.ruff.lint.mccabe]
max-complexity = 24

[tool.ruff.lint.pydocstyle]
convention = "numpy"



